Simplify ZSH configuration
Instead of computing the host color at each startup, write a
configuration file with the actual value.  Since we use this repository
to manage the code on all nodes, it's a good way to avoid copying the
generated file everywhere and preer using Bolt to setup nodes according
to my needs.